requirements/doc.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
alabaster==0.7.13
    # via sphinx
babel==2.17.0
    # via sphinx
beautifulsoup4==4.13.4
    # via pydata-sphinx-theme
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
commonmark==0.9.1
    # via recommonmark
docutils==0.17.1
    # via
    #   pydata-sphinx-theme
    #   recommonmark
    #   sphinx
    #   sphinx-panels
idna==3.10
    # via requests
imagesize==1.4.1
    # via sphinx
jinja2==3.1.6
    # via sphinx
markupsafe==2.1.5
    # via jinja2
packaging==25.0
    # via
    #   pydata-sphinx-theme
    #   sphinx
pydata-sphinx-theme==0.8.1
    # via sphinx-book-theme
pygments==2.19.1
    # via sphinx
pytz==2025.2
    # via babel
pyyaml==6.0.2
    # via sphinx-book-theme
recommonmark==0.6.0
    # via -r requirements/doc.in
requests==2.32.4
    # via
    #   sphinx
    #   sphinxcontrib-images
    #   sphinxcontrib-youtube
snowballstemmer==3.0.1
    # via sphinx
soupsieve==2.7
    # via beautifulsoup4
sphinx==4.2.0
    # via
    #   -r requirements/doc.in
    #   pydata-sphinx-theme
    #   recommonmark
    #   sphinx-book-theme
    #   sphinx-copybutton
    #   sphinx-panels
    #   sphinxcontrib-contentui
    #   sphinxcontrib-images
    #   sphinxcontrib-youtube
sphinx-book-theme==0.3.3
    # via -r requirements/doc.in
sphinx-copybutton==0.5.0
    # via -r requirements/doc.in
sphinx-panels==0.6.0
    # via -r requirements/doc.in
sphinxcontrib-applehelp==1.0.4
    # via
    #   -r requirements/doc.in
    #   sphinx
sphinxcontrib-contentui==0.2.5
    # via -r requirements/doc.in
sphinxcontrib-devhelp==1.0.2
    # via
    #   -r requirements/doc.in
    #   sphinx
sphinxcontrib-htmlhelp==2.0.1
    # via
    #   -r requirements/doc.in
    #   sphinx
sphinxcontrib-images==0.9.4
    # via -r requirements/doc.in
sphinxcontrib-jsmath==1.0.1
    # via
    #   -r requirements/doc.in
    #   sphinx
sphinxcontrib-mermaid==0.7.1
    # via -r requirements/doc.in
sphinxcontrib-qthelp==1.0.3
    # via
    #   -r requirements/doc.in
    #   sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via
    #   -r requirements/doc.in
    #   sphinx
sphinxcontrib-youtube==1.3.0
    # via -r requirements/doc.in
typing-extensions==4.13.2
    # via beautifulsoup4
urllib3==2.2.3
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
